feat(INV-38): M2-T2 候选筛选器 — Screener + configurable rules
Add Phase 2 screener (大聪明 output stage):
- screener_rules in settings.py: configurable rule list via .env, format "factor_name:op:threshold"
  Default rules: roe_latest>=0.08, momentum_20d>=0.0, debt_ratio_latest<=0.70, volume_ratio_5_20>=0.80
- src/screener/screener.py: Screener.run() + ScreenerCandidate/ScreenerResult dataclasses
  Skip semantics: missing factor skips rule (no fail); symbol needs ≥1 matched rule to qualify
- src/screener/report.py: result_to_json serializer
- CLI: python -m src.cli screener run --asof DATE --out FILE
- Fix: DerivedFactor.id BigInteger→Integer for SQLite/test compatibility
- 18 new unit tests covering: parse_rule, pass/fail/skip/PIT/json (42 total, all green)

+++ alembic/versions/20260302_0002_add_derived_factors.py
@@ -17,7 +17,7 @@ depends_on = None
 def upgrade() -> None:
     op.create_table(
         'derived_factors',
-        sa.Column('id', sa.BigInteger(), primary_key=True, autoincrement=True),
+        sa.Column('id', sa.Integer(), primary_key=True, autoincrement=True),
         sa.Column('symbol', sa.String(length=32), nullable=False),
         sa.Column('market', sa.String(length=16), nullable=False),
         sa.Column('asof_date', sa.Date(), nullable=False),
